* get_memory_map(): Changed parameters types to fixed-width types. * Added BeOS compatibility wrappers for get_memory_map(), map_physical_memory(),   and create_area().
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@37232 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ headers/os/drivers/KernelExport.h
@@ -152,8 +152,8 @@ extern status_t		unlock_memory(void *buffer, size_t numBytes, uint32 flags);
 extern status_t		get_memory_map_etc(team_id team, const void *address,
 						size_t numBytes, physical_entry *table,
 						uint32* _numEntries);
-extern long			get_memory_map(const void *buffer, ulong size,
-						physical_entry *table, long numEntries);
+extern int32		get_memory_map(const void *buffer, size_t size,
+						physical_entry *table, int32 numEntries);
 extern area_id		map_physical_memory(const char *areaName,
 						phys_addr_t physicalAddress, size_t size, uint32 flags,
 						uint32 protection, void **_mappedAddress);
